Added useAuth hook
